feat: 🚀 (issue-1) - use nginx as reverse proxy
Nginx was configured to be used as reverse proxy for application. Inside
docker-compose was added a Nginx image that has a configuration that
allows Nginx work as the reverse proxy for the application. When
application it's accessed by it's default port Nginx will redirect to
the proxy defined inside nginx.conf.

BREAKING CHANGE: 🧨 Nginx as reverse proxy for the application. docker.compose.debug was
removed because now docker.compose.dev will serve as a unique compose
for both dev and debug, the enviroment value called
NPM_START_SCRIPT_SUFFIX it's responsible for conditionally run dev or
debug command.